+++ rollup.config.js
@@ -1,27 +1,7 @@
-/* eslint-disable */
+require('ts-node').register({
+  compilerOptions: {
+    module: 'CommonJS',
+  },
+});
 
-import commonConfig from './rollup.config.common';
-
-const output = process.env.DYNAMIC_IMPORT
-  ? [
-      {
-        dir: 'dist/es',
-        format: 'es',
-      },
-      {
-        dir: 'dist/cjs/',
-        format: 'cjs',
-      },
-    ]
-  : [
-      {
-        file: 'dist/module.js',
-        format: 'es',
-      },
-      {
-        file: 'dist/module.cjs.js',
-        format: 'cjs',
-      },
-    ];
-
-export default commonConfig(output, true);
+module.exports = require('./rollup.config.common.ts');