+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 ### 🏠 Internal
 - `e2e`
   - [#1555](https://github.com/wix-incubator/rich-content/pull/1555) stabalise flaky gallery tests
+- `general`
+  - [#1556](https://github.com/wix-incubator/rich-content/pull/1556) converted rollup to typescript
 
 </details>
 <hr/>